code/backend/requirements.txt
# Django core
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9

# Caching
django-redis==5.4.0

# Async support
asgiref==3.8.1
uvicorn[standard]==0.29.0

# Auth
djangorestframework-simplejwt==5.3.1

# Quant / Math
scipy==1.13.0
numpy==1.26.4
pandas==2.2.2
python-dateutil>=2.9.0

# Data feeds
httpx==0.27.0
yfinance==0.2.38

# Testing
pytest==8.2.0
pytest-django==4.8.0
pytest-asyncio==0.23.6
factory-boy==3.3.0

# Dev
python-dotenv==1.0.1
